fix: split diplomacy widgetbook stories into own part file (#3622)
catalog_panels.dart exceeded the 1000-line part-fragment limit
(repo.part_unit_size) after adding the Relative Power Line stories.
Move the Diplomacy Panel + Relative Power Line stories and their
helpers into catalog_diplomacy_panel.dart to restore the Quality gate.

Refs #3622
